[#5] Move qos encoding-decoding into qos_codec.hpp
